requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
timm>=0.9.12
wandb>=0.16.0
accelerate>=0.24.0
einops>=0.7.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
datasets>=2.16.0
bitsandbytes>=0.43.0
